Move `crate::values` to `crate::ops::constant`
