+++ quantinuum-hugr/src/ops/constant/custom.rs
@@ -1,7 +1,8 @@
-//! Representation of values (shared between [Const] and in future [TypeArg])
+//! Representation of custom constant values.
 //!
-//! [Const]: crate::ops::Const
-//! [TypeArg]: crate::types::type_param::TypeArg
+//! These can be used as [`Const`] operations in HUGRs.
+//!
+//! [`Const`]: crate::ops::Const
 
 use std::any::Any;
 
@@ -26,7 +27,7 @@ pub trait CustomConst:
     /// An identifier for the constant.
     fn name(&self) -> SmolStr;
 
-    /// The extension(s) defining the custom value
+    /// The extension(s) defining the custom constant
     /// (a set to allow, say, a [List] of [USize])
     ///
     /// [List]: crate::std_extensions::collections::LIST_TYPENAME
@@ -39,7 +40,8 @@ pub trait CustomConst:
     }
 
     /// Compare two constants for equality, using downcasting and comparing the definitions.
-    // Can't derive PartialEq for trait objects
+    ///
+    /// If the type implements `PartialEq`, use [`downcast_equal_consts`] to compare the values.
     fn equal_consts(&self, _other: &dyn CustomConst) -> bool {
         // false unless overloaded
         false
@@ -51,11 +53,11 @@ pub trait CustomConst:
 
 /// Const equality for types that have PartialEq
 pub fn downcast_equal_consts<T: CustomConst + PartialEq>(
-    value: &T,
+    constant: &T,
     other: &dyn CustomConst,
 ) -> bool {
     if let Some(other) = other.as_any().downcast_ref::<T>() {
-        value == other
+        constant == other
     } else {
         false
     }
@@ -65,7 +67,7 @@ impl_downcast!(CustomConst);
 impl_box_clone!(CustomConst, CustomConstBoxClone);
 
 #[derive(Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
-/// A value stored as a serialized blob that can report its own type.
+/// A constant value stored as a serialized blob that can report its own type.
 pub struct CustomSerialized {
     typ: Type,
     value: serde_yaml::Value,
@@ -113,41 +115,5 @@ impl CustomConst for CustomSerialized {
 impl PartialEq for dyn CustomConst {
     fn eq(&self, other: &Self) -> bool {
         (*self).equal_consts(other)
-    }
-}
-
-#[cfg(test)]
-pub(crate) mod test {
-    use super::*;
-    use crate::ops::Const;
-    use crate::std_extensions::arithmetic::float_types::{self, FLOAT64_TYPE};
-    use crate::types::CustomType;
-
-    #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-
-    /// A custom constant value used in testing
-    pub(crate) struct CustomTestValue(pub CustomType);
-    #[typetag::serde]
-    impl CustomConst for CustomTestValue {
-        fn name(&self) -> SmolStr {
-            format!("CustomTestValue({:?})", self.0).into()
-        }
-
-        fn extension_reqs(&self) -> ExtensionSet {
-            ExtensionSet::singleton(self.0.extension())
-        }
-
-        fn get_type(&self) -> Type {
-            self.0.clone().into()
-        }
-    }
-
-    pub(crate) fn serialized_float(f: f64) -> Const {
-        CustomSerialized {
-            typ: FLOAT64_TYPE,
-            value: serde_yaml::Value::Number(f.into()),
-            extensions: float_types::EXTENSION_ID.into(),
-        }
-        .into()
     }
 }
